First working prototype.

+++ code/proc/proc_posix.c
@@ -5,40 +5,51 @@
  *      Author: joao
  */
 
-#define _POSIX_SOURCE
-#include <sys/types.h>
 #include <signal.h>
 
-#include "../../code/proc/proc.h"
-#include "../../code/proc/proc_posix.h"
+#include <proc/proc.h>
+#include <proc/proc_posix.h>
+#include <error/error_proc.h>
 
 t_status proc_module_supported( void )
 {
-  return STATUS_SUCCESS;	// Posix supports process interface
+  RETURN_STATUS_SUCCESS;	// Posix supports process interface
 }
 
 
-t_status proc_pid_get( t_pid * p_pid )
+t_status proc_id_get( t_pid * p_pid )
 {
  t_status	st;
 
  status_reset( & st );
 
- *p_pid = (t_pid) getpid();
+ if( p_pid == (t_pid) 0 )
+     status_iset( OSAPI_MODULE_PROC, __func__, e_proc_params, &st );
+ else
+   {
+     *p_pid = (t_pid) getpid();
+     if( *p_pid == -1 )
+         status_iset( OSAPI_MODULE_PROC, __func__, errno, &st );
+   }
 
  return st;
 }
 
 
-t_status proc_parentPid_get( t_pid * p_pid )
+t_status proc_parentID_get( t_pid * p_pid )
 {
  t_status	st;
 
  status_reset( & st );
 
- // STATUS_SET( status_modules.PROC, status_func.PARENTPID_GET, );
-
- *p_pid = (t_pid) getppid();
+ if( p_pid == (t_pid) 0 )
+     status_iset( OSAPI_MODULE_PROC, __func__, e_proc_params, &st );
+ else
+   {
+     *p_pid = (t_pid) getppid();
+     if( *p_pid == -1 )
+         status_iset( OSAPI_MODULE_PROC, __func__, errno, &st );
+   }
 
  return st;
 }
@@ -50,8 +61,13 @@ t_status proc_signal_send( t_pid pid, t_signal sig )
 
  status_reset( & st );
 
- if( kill( pid, sig ) != 0 )
-     status_set( PROC, f_proc_signal_send, (unsigned int) errno, &st );
+ if( pid <= (t_pid) 0 || sig <= (t_signal) 0 )
+     status_iset( OSAPI_MODULE_PROC, __func__, e_proc_params, &st );
+ else
+   {
+     if( kill( pid, sig ) != 0 )
+         status_eset( OSAPI_MODULE_PROC, __func__, errno, &st );
+   }
 
  return st;
 }
